requirements.txt
# requirements.txt

attrs==23.1.0
cachetools==5.3.2
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
distlib==0.3.7
google-api-core==2.14.0
google-auth==2.23.4
google-cloud==0.34.0
google-cloud-bigquery==3.13.0
google-cloud-bigquery-storage==2.22.0
google-cloud-core==2.3.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
idna==3.6
isodate==0.6.1
lxml==4.9.3
more-itertools==10.1.0
numpy==1.26.2
packaging==23.2
pandas==2.1.3
pendulum==2.1.2
platformdirs==4.0.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pytzdata==2020.1
requests==2.31.0
requests-file==1.5.1
requests-toolbelt==1.0.0
rsa==4.9
salesforce-bulk==2.2.0
simple-salesforce==1.12.5
six==1.16.0
tzdata==2023.3
unicodecsv==0.14.1
urllib3==2.1.0
zeep==4.2.1
